requirements.txt
pdfplumber
tabula-py
pandas
requests
PyYAML
